tests: Refactor index.js in a way we can test the output
