+++ src/index.js
@@ -3,15 +3,11 @@
 
 'use strict'
 
-const sgf = require('staged-git-files')
 const appRoot = require('app-root-path')
-const Listr = require('listr')
 const cosmiconfig = require('cosmiconfig')
 
 const packageJson = require(appRoot.resolve('package.json')) // eslint-disable-line
-const runScript = require('./runScript')
-const generateTasks = require('./generateTasks')
-const resolveGitDir = require('./resolveGitDir')
+const runAll = require('./runAll')
 
 // Force colors for packages that depend on https://www.npmjs.com/package/supports-color
 // but do this only in TTY mode
@@ -26,68 +22,7 @@ cosmiconfig('lint-staged', {
     .then((result) => {
         // result.config is the parsed configuration object
         // result.filepath is the path to the config file that was found
-        const config = result.config
-        const verbose = config.verbose
-        // Output config in verbose mode
-        if (verbose) console.log(config)
-        const concurrent = typeof config.concurrent !== 'undefined' ? config.concurrent : true
-        const renderer = verbose ? 'verbose' : 'update'
-        const gitDir = resolveGitDir(config)
-        sgf.cwd = gitDir
-
-        sgf('ACM', (err, files) => {
-            if (err) {
-                console.error(err)
-            }
-
-            /* files is an Object{ filename: String, status: String } */
-            const filenames = files.map(file => file.filename)
-            const tasks = generateTasks(config, filenames)
-                .map(task => ({
-                    title: `Running tasks for ${ task.pattern }`,
-                    task: () => (
-                        new Listr(
-                            runScript(
-                                task.commands,
-                                task.fileList,
-                                packageJson,
-                                { gitDir, verbose }
-                            ), {
-                                // In sub-tasks we don't want to run concurrently
-                                // and we want to abort on errors
-                                concurrent: false,
-                                exitOnError: true
-                            }
-                        )
-                    ),
-                    skip: () => {
-                        if (task.fileList.length === 0) {
-                            return `No staged files match ${ task.pattern }`
-                        }
-                        return false
-                    }
-                }))
-
-
-            if (tasks.length) {
-                new Listr(tasks, {
-                    concurrent,
-                    renderer,
-                    exitOnError: !concurrent // Wait for all errors when running concurrently
-                })
-                    .run()
-                    .catch((error) => {
-                        if (Array.isArray(error.errors)) {
-                            error.errors.forEach((lintError) => {
-                                console.error(lintError.message)
-                            })
-                        } else {
-                            console.log(error.message)
-                        }
-                        process.exit(1)
-                    })
-            }
-        })
+        runAll(packageJson, result.config)
     })
     .catch((parsingError) => {
         console.error(`Could not parse lint-staged config.
