+++ src/runAll.js
@@ -1,0 +1,75 @@
+/* global process */
+/* eslint no-console: 0 */
+
+'use strict'
+
+const sgf = require('staged-git-files')
+const Listr = require('listr')
+const runScript = require('./runScript')
+const generateTasks = require('./generateTasks')
+const resolveGitDir = require('./resolveGitDir')
+const getConfig = require('./getConfig')
+
+module.exports = function (packageJson, config) {
+    const { gitDir, verbose, concurrent, renderer } = getConfig(config)
+    sgf.cwd = resolveGitDir(gitDir)
+    return new Promise((resolve, reject) => {
+        sgf('ACM', (err, files) => {
+            if (err) {
+                console.error(err)
+            }
+
+            /* files is an Object{ filename: String, status: String } */
+            const filenames = files.map(file => file.filename)
+            const tasks = generateTasks(config, filenames)
+                .map(task => ({
+                    title: `Running tasks for ${ task.pattern }`,
+                    task: () => (
+                        new Listr(
+                            runScript(
+                                task.commands,
+                                task.fileList,
+                                packageJson,
+                                { gitDir, verbose }
+                            ), {
+                                // In sub-tasks we don't want to run concurrently
+                                // and we want to abort on errors
+                                concurrent: false,
+                                exitOnError: true
+                            }
+                        )
+                    ),
+                    skip: () => {
+                        if (task.fileList.length === 0) {
+                            return `No staged files match ${ task.pattern }`
+                        }
+                        return false
+                    }
+                }))
+
+
+            if (tasks.length) {
+                new Listr(tasks, {
+                    concurrent,
+                    renderer,
+                    exitOnError: !concurrent // Wait for all errors when running concurrently
+                })
+                    .run()
+                    .then(() => {
+                        resolve()
+                    })
+                    .catch((error) => {
+                        if (Array.isArray(error.errors)) {
+                            error.errors.forEach((lintError) => {
+                                console.error(lintError.message)
+                            })
+                        } else {
+                            console.log(error.message)
+                        }
+                        process.exit(1)
+                        reject()
+                    })
+            }
+        })
+    })
+}
